fix(bq-export): cast nullable int cols to pandas Int64 to prevent FLOAT64/INT64 MERGE error
Root cause: pd.to_numeric(..., errors='coerce') returns float64 when any
value is NaN/None (standard Python int can't hold NaN). BigQuery auto-
detects the staging table columns as FLOAT64, but the canonical incidents
table has status_code, ai_confidence, heuristic_confidence, occurrence_count
typed as INT64. The MERGE fails with:
  'Value of type FLOAT64 cannot be assigned to status_code, which has type INT64'

Fix: chain .astype('Int64') after pd.to_numeric().
pandas nullable integer type (capital I) stores NaN as pd.NA, maps to
BQ INT64 on upload — no schema mismatch.

+++ backend/vanguard/export/incident_exporter.py
@@ -269,10 +269,14 @@ class IncidentExporter:
             if col in df.columns:
                 df[col] = pd.to_datetime(df[col], utc=True, errors="coerce")
         
-        # Ensure integer columns
+        # Ensure integer columns — use pandas nullable Int64 (capital I), NOT float64.
+        # pd.to_numeric() alone returns float64 when NaN is present (standard int can't hold NaN).
+        # BigQuery canonical table has these columns as INT64, so the MERGE rejects float64 values
+        # with "Value of type FLOAT64 cannot be assigned to <col>, which has type INT64".
+        # pandas nullable integer ("Int64") handles NaN correctly and maps to BQ INT64.
         for col in ["occurrence_count", "status_code", "ai_confidence", "heuristic_confidence"]:
             if col in df.columns:
-                df[col] = pd.to_numeric(df[col], errors="coerce")
+                df[col] = pd.to_numeric(df[col], errors="coerce").astype("Int64")
         
         # Log label distribution
         if "ml_label" in df.columns:
